import-q/pyproject.toml
[project]
name = "import-q"
version = "0.1.0"
description = "import questions from official BIA PDF, BIA-AF and annales-bia"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "diskcache>=5.6.3",
    "google-genai>=1.59.0",
    "levenshtein>=0.27.3",
    "python-dotenv>=1.0.1",
    "python-rapidjson>=1.23",
    "requests>=2.32.4",
    "selenium>=4.39.0",
    "sqlmodel>=0.0.29",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pyright>=1.1.408",
    "ruff>=0.14.11",
]
